+++ app/controllers/project_users_controller.rb
@@ -25,9 +25,11 @@ class ProjectUsersController < ApplicationController
     respond_to do |format|
       if @project and (not @user.blank? or not invite_email.blank?)
         if @user
-          @project_user = @project.project_users.where(user_id: @user.id).first_or_create( creator_id: current_user.id, editor: (params[:project_user][:editor] == 'true') )
+          @project_user = @project.project_users.where(user_id: @user.id).first_or_create( creator_id: current_user.id )
+          @project_user.update( editor: (params[:project_user][:editor] == 'true') )
         elsif not invite_email.blank?
-          @project_user = @project.project_users.where(invite_email: invite_email).first_or_create( creator_id: current_user.id, editor: (params[:project_user][:editor] == 'true') )
+          @project_user = @project.project_users.where(invite_email: invite_email).first_or_create( creator_id: current_user.id )
+          @project_user.update( editor: (params[:project_user][:editor] == 'true') )
           @project_user.generate_invite_token!
         end
         format.js { render 'index' }